fix for TB08, made tx month mandatory on smear form. Fixed issue in all patients enrolled list where Integer was being compared using ==

+++ web/src/org/openmrs/module/mdrtb/web/controller/reporting/PatientListContoller.java
@@ -289,11 +289,11 @@ public class PatientListContoller {
     		    	
     		    	if( tf.getAnatomicalSite()!=null) {
     		    		Integer asId = tf.getAnatomicalSite().getConceptId();
-    		    		if(asId==Context.getService(MdrtbService.class).getConcept(MdrtbConcepts.PULMONARY_TB).getConceptId()) {
+    		    		if(asId.intValue()==Context.getService(MdrtbService.class).getConcept(MdrtbConcepts.PULMONARY_TB).getConceptId().intValue()) {
     		    			report += openTD() + getMessage("mdrtb.lists.pulmonaryShort") + closeTD();
     		    		}
     		    		
-    		    		else if (asId==Context.getService(MdrtbService.class).getConcept(MdrtbConcepts.EXTRA_PULMONARY_TB).getConceptId()) {
+    		    		else if (asId.intValue()==Context.getService(MdrtbService.class).getConcept(MdrtbConcepts.EXTRA_PULMONARY_TB).getConceptId().intValue()) {
     		    			report += openTD() + getMessage("mdrtb.lists.extrapulmonaryShort") + closeTD();
     		    		}
